requirements.txt
python-telegram-bot==20.6
python-magic==0.4.27
requests==2.31.0